requirements.txt
asgiref==3.8.1
certifi==2024.7.4
dj-database-url==2.2.0
Django==4.2.15
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elasticsearch==7.9.1
elasticsearch-dsl==7.4.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
